Pull Request 템플릿 적용을 위한 PR

+++ src/main/java/GDG/whatssue/domain/file/service/impl/S3UploadService.java
@@ -1,6 +1,8 @@
 package GDG.whatssue.domain.file.service.impl;
 
+import GDG.whatssue.domain.file.entity.UploadFile;
 import GDG.whatssue.domain.file.service.FileUploadService;
+import GDG.whatssue.domain.file.FileConst;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
@@ -10,6 +12,7 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 @Service
 @RequiredArgsConstructor
@@ -23,21 +26,26 @@ public class S3UploadService implements FileUploadService {
     @Value("${cloud.aws.s3.bucket}")
     public String bucket;
 
-
     @Override
-    public String saveFile(MultipartFile multipartFile, String dirName) throws IOException {
-        String originalFileName = multipartFile.getOriginalFilename();
+    public UploadFile uploadFile(MultipartFile multipartFile, String dirName) throws IOException {
+        String storeFileName;
+        //저장 사진이 없으면 기본 경로 반환
+        if (multipartFile == null) {
+            storeFileName = dirName + FileConst.DEFAULT_IMAGE_NAME;
+            return UploadFile.of(storeFileName,storeFileName);
+        }
 
-        String fileName = getFileName(dirName, originalFileName);
+        String originalFileName = multipartFile.getOriginalFilename();
+        storeFileName = getFileName(dirName, originalFileName);
 
         ObjectMetadata metadata = new ObjectMetadata();
         metadata.setContentLength(multipartFile.getSize());
         metadata.setContentType(multipartFile.getContentType());
 
-        amazonS3.putObject(new PutObjectRequest(bucket, fileName, multipartFile.getInputStream(), metadata)
+        amazonS3.putObject(new PutObjectRequest(bucket, storeFileName, multipartFile.getInputStream(), metadata)
             .withCannedAcl(CannedAccessControlList.PublicRead));
 
-        return fileName;
+        return UploadFile.of(originalFileName, storeFileName);
     }
 
     @Override
@@ -48,12 +56,12 @@ public class S3UploadService implements FileUploadService {
 
     @Override
     public void deleteFile(String storeFileName) {
-        //TODO
-        amazonS3.deleteObject(bucket, storeFileName);
-    }
+        //default 파일이면 버킷에서 삭제하지 않는다
+        if (StringUtils.substringMatch(storeFileName, 0, "default")) {
+            return;
+        }
 
-    public String getFullPath(String fileName) {
-        return PATH + fileName;
+        amazonS3.deleteObject(bucket, storeFileName);
     }
 
     private String getFileName(String dirName, String originalFileName) {
